% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_builders.R
\name{build_biomarker_terms}
\alias{build_biomarker_terms}
\title{Build biomarker terms for  Cox models}
\usage{
build_biomarker_terms(
  biomarker,
  biomarker_type = c("baseline", "time_varying", "baseline_change"),
  baseline_suffix = "_bl",
  tv_suffix = "_tv",
  change_suffix = "_delta",
  interaction_var = NULL
)
}
\arguments{
\item{biomarker}{Biomarker base name.}

\item{biomarker_type}{One of "baseline", "time_varying",or "baseline_change".}

\item{baseline_suffix}{Suffix for baseline biomarker columns.}

\item{tv_suffix}{Suffix for time-varying biomarker columns.}

\item{change_suffix}{Suffix for change biomarker columns.}

\item{interaction_var}{Optional interaction variable name.}
}
\value{
A list with main_terms, other_terms, and rhs_terms.
}
\description{
Build biomarker terms for  Cox models
}
