% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_metadata.R
\name{construct_metadata}
\alias{construct_metadata}
\title{Build baseline or time varying survival metadata from long panel data}
\usage{
construct_metadata(
  metadata,
  id_col = "subject",
  time_col = "Visit",
  event_col = NULL,
  new_time_col = "time",
  new_event_col = "event",
  start_col = "start",
  stop_col = "stop",
  biomarker_type = c("baseline", "time_varying", "baseline_change")
)
}
\arguments{
\item{metadata}{A data frame in long format.}

\item{id_col}{Subject ID column name.}

\item{time_col}{Visit time column name.}

\item{event_col}{Event indicator column name coded 0/1.}

\item{new_time_col}{Visit time column name created for baseline analysis.}

\item{new_event_col}{Event indicator column name coded 0/1 created for baseline analysis.}

\item{start_col}{Start indicator column name}

\item{stop_col}{Stop indicator column name}

\item{biomarker_type}{One of "baseline", "time_varying",  or "baseline_change".}
}
\value{
A data frame with one row per subject.
}
\description{
For baseline survival analysis:
Time is set to the first event time if an event occurs, otherwise the last observed follow-up time.
For time varying survival analysis:
Converts visit-level longitudinal data into counting-process style
tart-stop intervals for time-varying Cox analysis.
}
