% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproccess_data.R
\name{preproccess_data}
\alias{preproccess_data}
\title{preproccess survival metadata, biomarker table and covariate table}
\usage{
preproccess_data(
  metadata,
  biomarkers,
  id_col = "subject",
  time_col = "Visit",
  new_time_col = "time",
  start_col = "start",
  event_col = NULL,
  new_event_col = "event",
  stop_col = "stop",
  baseline_feature_table = NULL,
  time_varying_feature_table = NULL,
  covariates_table = NULL,
  biomarker_type = c("baseline", "time_varying", "baseline_change"),
  baseline_suffix = "_bl",
  tv_suffix = "_tv",
  change_suffix = "_delta"
)
}
\arguments{
\item{metadata}{Survival metadata,post formatting.}

\item{biomarkers}{Vector of Biomarker base names.}

\item{id_col}{Subject ID column name}

\item{time_col}{Time column name}

\item{new_time_col}{Visit time column name created for baseline analysis.}

\item{start_col}{Start column name}

\item{event_col}{event column name}

\item{new_event_col}{Event indicator column name coded 0/1 created for baseline analysis.}

\item{stop_col}{Stop column name}

\item{baseline_feature_table}{Optional data input for baseline biomarker measurements.}

\item{time_varying_feature_table}{Optional data input for longitudinal  biomarker measurements.baseline or time varying, need to input either one or both.}

\item{covariates_table}{Optional data input for covariates  measurements.}

\item{biomarker_type}{Biomarker mode.}

\item{baseline_suffix}{Baseline suffix.}

\item{tv_suffix}{Time-varying suffix.}

\item{change_suffix}{Change suffix.}
}
\value{
A data frame with survival metadata and biomarker measurements
}
\description{
preproccess survival metadata, biomarker table and covariate table
}
