% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_single_feature_cox.R
\name{run_single_cox_flexible}
\alias{run_single_cox_flexible}
\title{Fit one flexible Cox model for one biomarker}
\usage{
run_single_cox_flexible(
  data,
  biomarker,
  event_col,
  biomarker_type = c("baseline", "time_varying", "baseline_change"),
  time_col = NULL,
  start_col = NULL,
  stop_col = NULL,
  covariates = NULL,
  interaction_var = NULL,
  time_varying_coefficients = FALSE,
  center_time = 0,
  baseline_suffix = "_bl",
  tv_suffix = "_tv",
  change_suffix = "_delta",
  ties = "efron"
)
}
\arguments{
\item{data}{Input data frame.}

\item{biomarker}{Biomarker base name.}

\item{event_col}{Event column name.}

\item{biomarker_type}{Biomarker mode.}

\item{time_col}{Time column for standard Cox.}

\item{start_col}{Start column for start-stop Cox.}

\item{stop_col}{Stop column for start-stop Cox.}

\item{covariates}{Optional covariate names.}

\item{interaction_var}{Optional interaction variable name.}

\item{time_varying_coefficients}{Whether to add a time-varying coefficient
for a baseline biomarker.}

\item{center_time}{Center time if needed}

\item{baseline_suffix}{Baseline suffix.}

\item{tv_suffix}{Time-varying suffix.}

\item{change_suffix}{Change suffix.}

\item{ties}{Tie handling method.}
}
\value{
A data frame of model output.
}
\description{
Fit one flexible Cox model for one biomarker
}
