% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_surv_formula.R
\name{build_surv_formula}
\alias{build_surv_formula}
\title{Build survival formula}
\usage{
build_surv_formula(
  event_col,
  time_col = NULL,
  start_col = NULL,
  stop_col = NULL,
  rhs_terms
)
}
\arguments{
\item{event_col}{Event column name.}

\item{time_col}{Time column name for standard Cox model.}

\item{start_col}{Start column name for start-stop Cox model.}

\item{stop_col}{Stop column name for start-stop Cox model.}

\item{rhs_terms}{Right-hand side formula terms.}
}
\value{
A formula object.
}
\description{
Build survival formula
}

\details{
Column names are represented as symbols, supporting spaces and punctuation.
Interval columns must be supplied together.
}
