% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_FDR.R
\name{calculate_FDR}
\alias{calculate_FDR}
\title{calculate FDRs}
\usage{
calculate_FDR(
  res,
  terms,
  effect_col = "effect_type",
  term_col = "term",
  pvalue_col = "pvalue",
  main_term = "main_terms",
  other_terms = "other_terms",
  other_terms_vector = c("other_terms", "covariate")
)
}
\arguments{
\item{res}{output from multiple cox models}

\item{terms}{term_info from term builders}

\item{effect_col}{effect_type column from res.}

\item{term_col}{Term column from res.}

\item{pvalue_col}{P value column from res}

\item{main_term}{Main terms from term info}

\item{other_terms}{Other terms from term info}

\item{other_terms_vector}{Other_terms or covariate in the effect_type column}
}
\value{
A data frame with FDR calculated per term
}
\description{
calculate FDRs
}
